tests/wpt/web-platform-tests/resource-timing/document-domain-no-impact-loader.sub.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
    const t = async_test("Finite resource timing entries buffer size");
    addEventListener("message", t.step_func_done(e => {
        assert_equals(e.data, "PASS", "Document domain had no impact on the timing-allow check");
    }));
window.open("//{{domains[www]}}:{{ports[http][1]}}/resource-timing/resources/document-domain-no-impact.sub.html");
</script>
