Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/resource-timing/document-domain-no-impact-loader.sub-expected.txt

CONSOLE MESSAGE: SecurityError: Attempted to use a non-registrable domain.

FAIL Finite resource timing entries buffer size assert_equals: Document domain had no impact on the timing-allow check expected "PASS" but got "FAIL - iframe document.domain was not set"
